Added better index for snapshot_date

+++ db/migrate/20250414122147_add_snapshot_to_fund.rb
@@ -29,7 +29,7 @@ class AddSnapshotToFund < ActiveRecord::Migration[8.0]
 
     TABLES.each do |table|
       # Drop the table
-      drop_table "#{table.singularize}_snapshots"
+      drop_table "#{table.singularize}_snapshots" if ActiveRecord::Base.connection.table_exists?("#{table.singularize}_snapshots")
     end
   end
